Rename regression scripts.
The regression scripts are bash-tcl hybrid scripts.  The .sh extension
was casing some validators to validate it as a shell script but raise an
error when it encountered the tcl code portion of the script.  Renaming
the script to .tcl should validate the script as a tcl script and
validate it as a valid tcl script since the shell script portion is
embedded as string within valid tcl code.